+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.0] - 2024-04-12
+
+### Added
+
+- Added new plugin prameter:
+  - `allowMockLocations` - to enable mock locations for testing
+  - `foregroundNotificationTitle`, `foregroundNotificationText` - to configure foreground notification on Android
+
+## [4.1.0] - 2024-01-31
+
+### Added
+
+- Added `proxyDevicePushTokenCall` plugin property to enable calling HyperTrack SDK as soon as the device token is received in the AppDelegate
+
 ## [4.0.1] - 2023-11-11
 
 ### Fixed